Extract InMemorySessionFsSqliteHandler into reusable file
Move the in-memory SessionFsProvider + ISessionFsSqliteProvider test
implementation and SqliteCall record into a separate file for reuse.
